Make ALU great again! (2/3): Migrated Visual Studio solution to VS2017 and updated SDL2 to version 2.0.5 for all systems. Checked and working so far.

+++ Code/Data/Primitives.h
@@ -63,7 +63,7 @@ struct Float3 {
       return Float3(X * val, Y * val, Z * val);
     }
     float Distance(Float3 other) {
-      return sqrt(
+      return sqrtf(
         (X - other.X)*(X - other.X) +
         (Y - other.Y)*(Y - other.Y) +
         (Z - other.Z)*(Z - other.Z));
